# Conflicts:
#	gradle.properties
#	modchanges.md
#	src/generated/resources/.cache/59eb3dbb5f86130e09b3c62d89b9525ee01cf52d
#	src/generated/resources/.cache/8c466ff74522a93ac5a99c5384c18a642b921486
#	src/generated/resources/.cache/b7b3020c29f0ef8af7fcde9247b7030b8b16f43c
#	src/main/java/com/cmdpro/datanessence/client/ClientModEvents.java
#	src/main/java/com/cmdpro/datanessence/datagen/loot/ModBlockLootTables.java
#	src/main/java/com/cmdpro/datanessence/registry/BlockEntityRegistry.java

+++ src/main/java/com/cmdpro/datanessence/item/equipment/HammerAndChisel.java
@@ -1,19 +1,23 @@
 package com.cmdpro.datanessence.item.equipment;
 
+import com.cmdpro.datanessence.block.technical.StructureProtectorBlockEntity;
+import com.cmdpro.datanessence.registry.AttachmentTypeRegistry;
 import com.cmdpro.datanessence.registry.SoundRegistry;
 import com.cmdpro.datanessence.registry.TagRegistry;
 import net.minecraft.core.BlockPos;
-import net.minecraft.sounds.SoundEvents;
 import net.minecraft.sounds.SoundSource;
+import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
+import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.phys.AABB;
 
-import java.util.function.Consumer;
+import java.util.List;
 
 public class HammerAndChisel extends Item {
 
@@ -34,12 +38,25 @@ public class HammerAndChisel extends Item {
         return false;
     }
 
+    public boolean isPositionStructureProtected(Level world, BlockPos pos) {
+        if ( world.hasData(AttachmentTypeRegistry.STRUCTURE_CONTROLLERS) ) {
+            List<StructureProtectorBlockEntity> protectors = world.getData(AttachmentTypeRegistry.STRUCTURE_CONTROLLERS);
+            for (StructureProtectorBlockEntity thisProtector : protectors) {
+                AABB aabb = AABB.encapsulatingFullBlocks(thisProtector.getCorner1(), thisProtector.getCorner2());
+                if (aabb.contains(pos.getCenter())) {
+                    return true;
+                }
+            }
+        }
+        return false;
+    }
+
     // Destroy this block with the hammer.
-    public boolean hammerBlock(Level world, BlockPos target, ItemStack stack, Player player) {
+    public boolean hammerBlock(Level world, BlockPos target, ItemStack stack, Player player, EquipmentSlot hand) {
         BlockState targetState = world.getBlockState(target);
 
-        if (targetState.is(TagRegistry.Blocks.HAMMER_AND_CHISEL_COLLECTABLE)) {
-            stack.hurtAndBreak(1, player, player.getEquipmentSlotForItem(stack));
+        if (targetState.is(TagRegistry.Blocks.HAMMER_AND_CHISEL_COLLECTABLE) && !isPositionStructureProtected(world, target)) {
+            stack.hurtAndBreak(1, player, hand);
             world.destroyBlock(target, true);
             return true;
         }
@@ -55,12 +72,13 @@ public class HammerAndChisel extends Item {
         Level world = context.getLevel();
         BlockPos pos = context.getClickedPos();
         Player player = context.getPlayer();
+        var hand = context.getHand() == InteractionHand.MAIN_HAND ? EquipmentSlot.MAINHAND : EquipmentSlot.OFFHAND;
 
         if (world.isClientSide())
             return InteractionResult.SUCCESS;
 
         // TODO make it so you need to click twice on the same block to destroy it, rather than just once.
-        if (hammerBlock(world, pos, context.getItemInHand(), player)) {
+        if (hammerBlock(world, pos, context.getItemInHand(), player, hand)) {
             playSound(world, pos);
             return InteractionResult.CONSUME;
         }
